requirements.txt
connexion>=2, <3
connexion[swagger-ui]>=2, <3
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
typing >= 3.5.2.2
pyyaml < 5.3; python_version<="2.7"
setuptools >= 21.0.0
jinja2==3.0.3
psycopg2-binary
bcrypt~=4.1.3
Flask-JWT-Extended

Flask~=2.2.5
six~=1.16.0
PyJWT~=2.8.0